common/gui/modularui2/cover -> common/covers/gui
