+++ src/main/java/gregtech/common/covers/CoverControlsWork.java
@@ -20,8 +20,8 @@ import gregtech.api.interfaces.tileentity.ICoverable;
 import gregtech.api.interfaces.tileentity.IMachineProgress;
 import gregtech.api.util.GTUtility;
 import gregtech.common.covers.conditions.RedstoneCondition;
-import gregtech.common.gui.modularui2.cover.CoverControlsWorkGui;
-import gregtech.common.gui.modularui2.cover.CoverGui;
+import gregtech.common.covers.gui.CoverControlsWorkGui;
+import gregtech.common.covers.gui.CoverGui;
 import gregtech.common.gui.mui1.cover.ControlsWorkUIFactory;
 
 public class CoverControlsWork extends CoverLegacyData {
