+++ src/main/java/gregtech/common/covers/CoverFacadeBase.java
@@ -27,8 +27,8 @@ import gregtech.api.render.TextureFactory;
 import gregtech.api.util.GTByteBuffer;
 import gregtech.api.util.GTRenderingWorld;
 import gregtech.api.util.GTUtility;
-import gregtech.common.gui.modularui2.cover.CoverFacadeBaseGui;
-import gregtech.common.gui.modularui2.cover.CoverGui;
+import gregtech.common.covers.gui.CoverFacadeBaseGui;
+import gregtech.common.covers.gui.CoverGui;
 import gregtech.common.gui.mui1.cover.FacadeBaseUIFactory;
 import io.netty.buffer.ByteBuf;
