+++ src/main/java/gregtech/common/covers/CoverPump.java
@@ -15,8 +15,8 @@ import gregtech.api.interfaces.ITexture;
 import gregtech.api.interfaces.tileentity.ICoverable;
 import gregtech.api.interfaces.tileentity.IMachineProgress;
 import gregtech.api.util.GTUtility;
-import gregtech.common.gui.modularui2.cover.CoverGui;
-import gregtech.common.gui.modularui2.cover.CoverIOBaseGui;
+import gregtech.common.covers.gui.CoverGui;
+import gregtech.common.covers.gui.CoverIOBaseGui;
 import gregtech.common.gui.mui1.cover.PumpUIFactory;
 
 public class CoverPump extends CoverIOBase {
